+++ src/record/record-handler.ts
@@ -305,6 +305,10 @@ export class RecordHandler {
     this.sendSetData(recordName, -1, args)
   }
 
+  public delete (recordName: string, callback?: (error: string | null) => void): void | Promise<void> {
+    // TODO: Use a delete service to make the logic in record core and here common
+  }
+
   private sendSetData (recordName: string, version: number, args: utils.RecordSetArguments): void {
     const { path, data, callback } = args
     if (!recordName || typeof recordName !== 'string' || recordName.length === 0) {
@@ -434,8 +438,6 @@ export class RecordHandler {
     if (!recordCore) {
       recordCore = new RecordCore(recordName, this.services, this.options, this.recordServices, this.removeRecord)
       this.recordCores.set(recordName, recordCore)
-    } else {
-      recordCore.usages++
     }
     return recordCore
   }
@@ -444,11 +446,13 @@ export class RecordHandler {
     this.dirtyService.whenLoaded(this, this._syncDirtyRecords)
   }
 
+  // TODO: Expose issues here, as there isn't a reason why a record core needs to exist in
+  // order to sync up
   private _syncDirtyRecords () {
     const dirtyRecords = this.dirtyService.getAll()
     for (const recordName in dirtyRecords) {
       const recordCore = this.recordCores.get(recordName)
-      if (recordCore && recordCore.usages > 0) {
+      if (recordCore && recordCore.references.size > 0) {
         // if it isn't zero.. problem.
         continue
       }