+++ src/storage/noop-storage-service.ts
@@ -2,6 +2,8 @@ import { RecordOfflineStore, offlineStoreWriteResponse } from '../client'
 import {RecordData} from '../../binary-protocol/src/message-constants'
 
 export class NoopStorage implements RecordOfflineStore {
+  public isReady = true
+
   public get (recordName: string, callback: ((recordName: string, version: number, data: RecordData) => void)) {
     setTimeout(callback.bind(this, recordName, -1, null), 0)
   }