+++ src/util/state-machine.ts
@@ -5,12 +5,14 @@ export class StateMachine {
 
   private transitions: any
   private context: any
+  private history: Array<any>
 
   constructor (logger: any, private stateMachine: any) {
     this.inEndState = false
     this.transitions = stateMachine.transitions
     this.state = stateMachine.init
     this.context = stateMachine.context
+    this.history = [{ oldState: '-', newState: this.state, transitionName: '-' }]
   }
 
   /**
@@ -21,6 +23,7 @@ export class StateMachine {
     for (let i = 0; i < this.transitions.length; i++) {
       transition = this.transitions[i]
       if (transitionName === transition.name && (this.state === transition.from || transition.from === undefined)) {
+        this.history.push({ oldState: this.state, transitionName, newState: transition.to })
         const oldState = this.state
         this.state = transition.to
         if (this.stateMachine.onStateChanged) {
@@ -33,7 +36,10 @@ export class StateMachine {
       }
     }
     const details = JSON.stringify({ transition: transitionName, state: this.state })
-    throw new Error(`Invalid state transition: ${details}`)
+    const debugHistory = this.history.reduce((result, entry) =>
+      result += `\n\tFrom ${entry.oldState} to ${entry.newState} via ${entry.transitionName}`
+    , '')
+    throw new Error(`Invalid state transition.\nDetails: ${details} \nHistory: ${debugHistory}`)
   }
 
 }